implicit number coercion for valueof and arrayify
