+++ src/transforms/dodge.js
@@ -1,6 +1,6 @@
 import IntervalTree from "interval-tree-1d";
 import {finite, positive} from "../defined.js";
-import {floatMap, maybeNamed, number} from "../options.js";
+import {identity, maybeNamed, number, valueof} from "../options.js";
 import {initializer} from "./basic.js";
 import {Position} from "../projection.js";
 
@@ -72,7 +72,7 @@ function dodge(y, x, anchor, padding, options) {
     if (!channels[x]) throw new Error(`missing channel: ${x}`);
     ({[x]: X} = Position(channels, scales, context));
     const r = R ? undefined : this.r !== undefined ? this.r : options.r !== undefined ? number(options.r) : 3;
-    if (R) R = floatMap(R.value, scales[R.scale]);
+    if (R) R = valueof(R.value, scales[R.scale] || identity, Float64Array);
     let [ky, ty] = anchor(dimensions);
     const compare = ky ? compareAscending : compareSymmetric;
     const Y = new Float64Array(X.length);
